muisque app
The project titled “MUSIQUE” of Music library management system is
management software for monitoring and controlling the features of
music.
Musique has to sale various musical items and each item comes with
number of models and varieties
The main objective of our project IS to make a free website which would
allow the users to listen to free music and even add their own collection
of songs or albums and contribute towards the broadening of the
website, which we were able to provide through this project.
The USER has the privilege to post/publish and maintain albums of their
own, each album having corresponding songs. In addition to these
features, the user can create and maintain playlists of songs. They even
have the ability to add song to favorites